+++ src/EventListener/Contao/LoadDataContainerListener.php
@@ -9,6 +9,9 @@
 namespace HeimrichHannot\IsotopeSubscriptionsBundle\EventListener\Contao;
 
 use Contao\CoreBundle\ServiceAnnotation\Hook;
+use Contao\System;
+use HeimrichHannot\IsotopeSubscriptionsBundle\Controller\FrontendModule\IsoActivationModuleController;
+use HeimrichHannot\IsotopeSubscriptionsBundle\Controller\FrontendModule\IsoCancellationModuleController;
 use HeimrichHannot\IsotopeSubscriptionsBundle\Manager\SubscriptionManager;
 
 /**
@@ -44,70 +47,14 @@ class LoadDataContainerListener
                  * Fields.
                  */
                 $fields = [
-                    'iso_addSubscription' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_addSubscription'],
-                        'exclude' => true,
-                        'inputType' => 'checkbox',
-                        'eval' => ['tl_class' => 'w50 long', 'submitOnChange' => true],
-                        'sql' => "char(1) NOT NULL default ''",
-                    ],
-                    'iso_subscriptionArchive' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_subscriptionArchive'],
-                        'exclude' => true,
-                        'inputType' => 'select',
-                        'foreignKey' => 'tl_iso_subscription_archive.title',
-                        'eval' => ['tl_class' => 'w50', 'mandatory' => true],
-                        'sql' => "int(10) unsigned NOT NULL default '0'",
-                    ],
-                    'iso_cancellationArchives' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_cancellationArchives'],
-                        'exclude' => true,
-                        'inputType' => 'select',
-                        'foreignKey' => 'tl_iso_subscription_archive.title',
-                        'eval' => ['tl_class' => 'w50', 'mandatory' => true, 'multiple' => true, 'chosen' => true],
-                        'sql' => 'blob NULL',
-                    ],
-                    'iso_addActivation' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_addActivation'],
-                        'exclude' => true,
-                        'inputType' => 'checkbox',
-                        'eval' => ['tl_class' => 'w50 clr', 'submitOnChange' => true],
-                        'sql' => "char(1) NOT NULL default ''",
-                    ],
-                    'iso_activationJumpTo' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_activationJumpTo'],
-                        'exclude' => true,
-                        'inputType' => 'pageTree',
-                        'foreignKey' => 'tl_page.title',
-                        'eval' => ['fieldType' => 'radio', 'tl_class' => 'w50'],
-                        'sql' => 'int(10) unsigned NOT NULL default 0',
-                        'relation' => ['type' => 'hasOne', 'load' => 'lazy'],
-                    ],
-                    'iso_checkForExitingSubscription' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_checkForExitingSubscription'],
-                        'exclude' => true,
-                        'filter' => true,
-                        'inputType' => 'checkbox',
-                        'eval' => ['tl_class' => 'w50'],
-                        'sql' => "char(1) NOT NULL default ''",
-                    ],
-                    'iso_activationNotification' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_activationNotification'],
-                        'exclude' => true,
-                        'inputType' => 'select',
-                        'options_callback' => ['NotificationCenter\tl_module', 'getNotificationChoices'],
-                        'eval' => ['includeBlankOption' => true, 'chosen' => true, 'tl_class' => 'w50', 'mandatory' => true],
-                        'sql' => "int(10) unsigned NOT NULL default '0'",
-                        'relation' => ['type' => 'hasOne', 'load' => 'lazy', 'table' => 'tl_nc_notification'],
-                    ],
-                    'iso_addSubscriptionCheckbox' => [
-                        'label' => &$GLOBALS['TL_LANG']['tl_module']['iso_addSubscriptionCheckbox'],
-                        'exclude' => true,
-                        'filter' => true,
-                        'inputType' => 'checkbox',
-                        'eval' => ['tl_class' => 'w50'],
-                        'sql' => "char(1) NOT NULL default ''",
-                    ],
+                    'iso_addSubscription' => $dca['fields']['iso_addSubscription'],
+                    'iso_subscriptionArchive' => $dca['fields']['iso_subscriptionArchive'],
+                    'iso_cancellationArchive' => $dca['fields']['iso_cancellationArchive'],
+                    'iso_addActivation' => $dca['fields']['iso_addActivation'],
+                    'iso_activationJumpTo' => $dca['fields']['iso_activationJumpTo'],
+                    'iso_checkForExitingSubscription' => $dca['fields']['iso_checkForExitingSubscription'],
+                    'iso_activationNotification' => $dca['fields']['iso_activationNotification'],
+                    'iso_addSubscriptionCheckbox' => $dca['fields']['iso_addSubscriptionCheckbox'],
                 ];
 
                 foreach (['iso_direct_checkout_products', 'iso_direct_checkout_product_types'] as $field) {
@@ -120,6 +67,39 @@ class LoadDataContainerListener
 
                     $dca['fields'][$field]['eval']['multiColumnEditor']['subpalettes']['iso_addSubscription'] = 'iso_subscriptionArchive,iso_addSubscriptionCheckbox,iso_addActivation';
                     $dca['fields'][$field]['eval']['multiColumnEditor']['subpalettes']['iso_addActivation'] = 'iso_activationNotification,iso_activationJumpTo';
+                }
+
+                // privacy
+                if (class_exists('\HeimrichHannot\PrivacyBundle\HeimrichHannotPrivacyBundle')) {
+                    $protocolManager = System::getContainer()->get(\HeimrichHannot\PrivacyBundle\Manager\ProtocolManager::class);
+
+                    $dca['fields']['iso_privacyEntryConfig'] = $protocolManager->getConfigFieldDca();
+                    $dca['fields']['iso_secondPrivacyEntryConfig'] = $protocolManager->getConfigFieldDca();
+
+                    // cancel
+                    $dca['palettes'][IsoCancellationModuleController::TYPE] = str_replace('iso_cancellationArchive',
+                        'iso_cancellationArchive,iso_privacyEntryConfig', $dca['palettes'][IsoCancellationModuleController::TYPE]);
+
+                    $dca['subpalettes']['iso_addActivation'] = str_replace('iso_activationLinkSentJumpTo',
+                        'iso_activationLinkSentJumpTo,iso_secondPrivacyEntryConfig', $dca['subpalettes']['iso_addActivation']);
+
+                    // activate
+                    $dca['palettes'][IsoActivationModuleController::TYPE] = str_replace('jumpTo',
+                        'jumpTo,iso_privacyEntryConfig', $dca['palettes'][IsoActivationModuleController::TYPE]);
+
+                    $fields = [
+                        'iso_privacyEntryConfig' => $dca['fields']['iso_privacyEntryConfig'],
+                    ];
+
+                    foreach (['iso_direct_checkout_products', 'iso_direct_checkout_product_types'] as $field) {
+                        $dca['fields'][$field]['eval']['multiColumnEditor']['fields'] = array_merge($dca['fields'][$field]['eval']['multiColumnEditor']['fields'], $fields);
+
+                        $dca['fields'][$field]['eval']['multiColumnEditor']['subpalettes']['iso_addSubscription'] = str_replace(
+                            'iso_addActivation',
+                            'iso_privacyEntryConfig,iso_addActivation',
+                            $dca['fields'][$field]['eval']['multiColumnEditor']['subpalettes']['iso_addSubscription']
+                        );
+                    }
                 }
 
                 break;