+++ src/Resources/contao/languages/de/default.php
@@ -5,9 +5,9 @@ $lang = &$GLOBALS['TL_LANG']['MSC'];
 /**
  * Messages
  */
-$lang['iso_subscriptionAlreadyExists']          = 'Unter der E-Mail-Adresse <em>%s</em> gibt es bereits ein Abo für den gewünschten Artikel <em>%s</em>.';
-$lang['iso_subscriptionDoesNotExist']           = 'Unter der E-Mail-Adresse <em>%s</em> gibt es kein Abo für den Artikel <em>%s</em>.';
-$lang['iso_subscriptionCancelledSuccessfully']  = 'Das Abo für den Artikel <em>%s</em> unter der E-Mail-Adresse <em>%s</em> wurde erfolgreich gelöscht.';
+$lang['iso_subscriptionAlreadyExists']          = 'Unter der E-Mail-Adresse <em>%s</em> gibt es bereits ein Abo für den gewünschten Artikel.';
+$lang['iso_subscriptionDoesNotExist']           = 'Unter der E-Mail-Adresse <em>%s</em> gibt es kein Abo für den Artikel.';
+$lang['iso_subscriptionCancelledSuccessfully']  = 'Das Abo für den Artikel unter der E-Mail-Adresse <em>%s</em> wurde erfolgreich gelöscht.';
 $lang['iso_subscriptionsCancelledSuccessfully'] = 'Die verfügbaren Abos für die E-Mail-Adresse <em>%s</em> wurden erfolgreich gelöscht.';
 $lang['iso_subscriptionActivatedSuccessfully']  = 'Ihr Abo wurde erfolgreich aktiviert.';
 $lang['iso_subscriptionAlreadyActivated']       = 'Das Abo wurde bereits aktiviert.';