+++ src/Resources/contao/languages/de/tl_module.php
@@ -5,12 +5,13 @@ $lang = &$GLOBALS['TL_LANG']['tl_module'];
 /**
  * Fields
  */
-$lang['iso_addSubscription']         = ['Abo hinzufügen', 'Wählen Sie diese Option, wenn Sie Abos für dieses Modul konfigurieren möchten.'];
-$lang['iso_subscriptionArchive']     = ['Abo-Archiv', 'Wählen Sie hier das Abo-Archiv aus, in dem bei Bestellung ein Abo angelegt werden soll. Es erfolgt eine 1:1-Abbildung der Feldernamen der Bestellung auf die Felder des Abos. Die Abo-Felder können in den Einstellungen konfiguriert werden.'];
-$lang['iso_addSubscriptionCheckbox'] = ['Checkbox für das Abo hinzufügen', 'Wählen Sie diese Option, wenn der Nutzer die Möglichkeit haben soll, das Abo optional zu aktivieren.'];
-$lang['iso_addActivation']           = ['Aktivierung hinzufügen', 'Wählen Sie diese Option, wenn das Abo erst per E-Mail bestätigt werden muss. Anderenfalls ist das Abo sofort aktiv.'];
-$lang['iso_activationNotification']  = ['Benachrichtigung (Aktivierung)', 'Wählen Sie hier die Benachrichtigung aus, die den Aktivierungslink enthält.'];
-$lang['iso_activationJumpTo']        = ['Weiterleitungsseite (Aktivierung)', 'Wählen Sie hier die Seite aus, zu der der Aktivierungslink weiterleiten soll. Die Seite muss ein Aktivierungsmodul enthalten.'];
-$lang['iso_cancellationArchives']    = ['Zu kündigende Abo-Archive', 'Wählen Sie hier die Abo-Archive aus, die beim Abschicken des Forumlars gekündigt werden sollen.'];
-$lang['email']                       = ['E-Mail-Adresse', 'Geben Sie hier eine E-Mail-Adresse ein.'];
-$lang['iso_change_address_archive']  = ['Abo-Archiv', 'Wählen Sie hier das Abo-Archiv aus, auf das die Adressänderung angewandt werden soll.'];
+$lang['iso_addSubscription']          = ['Abo hinzufügen', 'Wählen Sie diese Option, wenn Sie Abos für dieses Modul konfigurieren möchten.'];
+$lang['iso_subscriptionArchive']      = ['Abo-Archiv', 'Wählen Sie hier das Abo-Archiv aus, in dem bei Bestellung ein Abo angelegt werden soll. Es erfolgt eine 1:1-Abbildung der Feldernamen der Bestellung auf die Felder des Abos. Die Abo-Felder können in den Einstellungen konfiguriert werden.'];
+$lang['iso_addSubscriptionCheckbox']  = ['Checkbox für das Abo hinzufügen', 'Wählen Sie diese Option, wenn der Nutzer die Möglichkeit haben soll, das Abo optional zu aktivieren.'];
+$lang['iso_addActivation']            = ['Aktivierung hinzufügen', 'Wählen Sie diese Option, wenn das Abo erst per E-Mail bestätigt werden muss. Anderenfalls ist das Abo sofort aktiv.'];
+$lang['iso_activationNotification']   = ['Benachrichtigung (Aktivierung)', 'Wählen Sie hier die Benachrichtigung aus, die den Aktivierungslink enthält.'];
+$lang['iso_activationJumpTo']         = ['Weiterleitungsseite (Aktivierung)', 'Wählen Sie hier die Seite aus, zu der der Aktivierungslink weiterleiten soll. Die Seite muss ein Aktivierungsmodul enthalten.'];
+$lang['iso_activationLinkSentJumpTo'] = ['Weiterleitungsseite (Link gesendet)', 'Wählen Sie hier die Seite aus, zu weitergeleitet werden soll, nachdem der Aktivierungs-Benachrichtigung verschickt wurde.'];
+$lang['iso_cancellationArchive']      = ['Zu kündigendes Abo-Archiv', 'Wählen Sie hier das Abo-Archiv aus, das beim Abschicken des Formulars gekündigt werden soll.'];
+$lang['email']                        = ['E-Mail-Adresse', 'Geben Sie hier eine E-Mail-Adresse ein.'];
+$lang['iso_change_address_archive']   = ['Abo-Archiv', 'Wählen Sie hier das Abo-Archiv aus, auf das die Adressänderung angewandt werden soll.'];