+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] - 2021-06-07
+
+- refactoring for contao 4.9
+- added support for heimrichhannot/contao-privacy-bundle
+- added second opt out for cancellation module
+- after cancellation the subscription is now deleted completely
+
 ## [1.0.4] - 2020-01-16
 
 - use of model